Fix padding bug in hexToByteString
If we have an odd number of hex chars we need to 0 pad to fill the last byte. That padding needs to
be in the 4 most significant bits.

For example, if we have the hex char 'a' (binary 1010), the byte needs to be 00001010, which means
padding the pair of hex values with a 0 in the  first char: [0,10] => 00001010.